// frontend/src/shared/api/types.gen.ts
export type ErrorResponse = {
    status?: number;
    message?: string;
    path?: string;
    errors?: {
        [key: string]: string;
    };
    timestamp?: string;
};

export type ReviewRequest = {
    rating?: number;
    text?: string;
};

export type ReviewDto = {
    id?: string;
    listingId?: string;
    author?: UserDto;
    rating?: number;
    text?: string;
    createdAt?: string;
    updatedAt?: string;
};

export type UserDto = {
    id?: string;
    username?: string;
    email?: string;
    role?: 'admin' | 'user';
    verified?: 'none' | 'pending' | 'verified' | 'rejected';
    createdAt?: string;
    blocked?: boolean;
    verificationReason?: string;
    blockReason?: string;
};
// ...
export type AuthResponse = {
    user?: UserDto;
    accessToken?: string;
};